api/requirements.txt
chromadb
fastapi
pypdf
requests
beautifulsoup4
uvicorn
pyyaml
google-generativeai
aiogram
pydantic
pydantic_settings
free-proxy
gradio
nltk
pytest
beautifulsoup4
python-dotenv
langchain-google-genai 
langchain-community 
langgraph 
pyngrok
langchain_core
nest-asyncio